+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 4.2.0 - 2021-7-12
+### Changed
+
+* Add feature flag `record_location_unstable` to expand record location support.
+  Be warned that this is accomplished by leaking the record location strings.
+
+## 4.1.0 - 2020-10-21
+### Changed
+
+* Require `slog` 2.4 or greater
+* Remove dependency `crossbeam`
+* Require `log` 0.4.11 or greater
+
 ## 4.0.0 - 2018-08-13
 ### Changed